requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
cmake==3.27.2
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
filelock==3.12.3
idna==3.4
install==1.3.5
Jinja2==3.1.2
lit==16.0.6
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
Pillow==10.0.0
postgres==4.0
psycopg==3.1.10
psycopg2-binary==2.9.7
psycopg2-pool==1.1
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
sympy==1.12
torch==2.0.1
torchvision==0.15.2
triton==2.0.0
typing_extensions==4.7.1
urllib3==2.0.4
matplotlib==3.4.3
beautifulsoup4==4.10.0
